Split out common sandbox test code
